notifier: reduce the idle time to 1 minute
the motivation here was to not have the update run while the user is
using the system so as to not disturb workflows. alas, the problem is
that idle time is a poor measure because the system may never be idle
for long enough or the user returns before the update is complete
rendering the idle timeout pointless to begin with.

accepting defeat here, we are bring this value way down.

should this become a problem we may need to investigate better solutions
depending on what the actual problem is (e.g. cgroup constraints could
help with load issues). remains to be seen if there will be problems
though.

BUG: 447245